Remove outdated tmcli and merge features
