+++ Docs/script_audit.txt
@@ -51,7 +51,6 @@ safecron.sh - keep
 send_daily_roi_summary.py - keep
 streamlit_dashboard.py - keep
 tag_roi_tracker.py - keep
-tmcli.py - keep
 train_model_v6.py - keep
 train_modelv7.py - keep
 upload_logs_to_s3.sh - keep
